refactor: divide curves file into two
